move extrapolation stuff to new file
